pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qwen-markel-tts"
version = "0.1.0"
description = "RTX 5090 megakernel-backed Qwen3-TTS talker decoder with Pipecat integration"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "accelerate>=0.26.0",
    "ninja",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "aiohttp>=3.9.0",
    "soundfile>=0.12.1",
    "librosa>=0.10.0",
    "numpy>=1.26.0",
    "loguru>=0.7.0",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black",
    "isort",
]

[project.optional-dependencies]
pipecat = [
    "pipecat-ai>=0.0.53",
]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
